# docs/OmnitruckApi_swagger.yaml
definitions:
  ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status_text:
        type: string
    type: object
  omnitruck.ArchList:
    additionalProperties:
      $ref: '#/definitions/omnitruck.PackageMetadata'
    type: object
  omnitruck.PackageList:
    additionalProperties:
      $ref: '#/definitions/omnitruck.PlatformVersionList'
    type: object
  omnitruck.PackageMetadata:
    properties:
      sha1:
        type: string
      sha256:
        type: string
      url:
        type: string
      version:
        type: string
    type: object
...
  omnitruck.PlatformVersionList:
    additionalProperties:
      $ref: '#/definitions/omnitruck.ArchList'
    type: object
info:
  contact: {}
  description: Licensed Omnitruck API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Licensed Omnitruck API
  version: "1.0"
paths:
  /{channel}/{product}/download:
    get:
      consumes:
      - application/json
      description: |-
        Get details for a particular package.
        The `ACCEPT` HTTP header with a value of `application/json` must be provided in the request for a JSON response to be returned
      parameters:
      - description: Channel
        enum:
        - current
        - stable
        in: path
        name: channel
        required: true
        type: string
      - description: Product
        example: chef
        in: path
        name: product
        required: true
        type: string
      - description: Platform, valid values are returned from the `/platforms` endpoint.
        example: ubuntu
        in: query
        name: p
        required: true
        type: string
      - description: 'Platform Version, possible values depend on the platform. For
          example, Ubuntu: 16.04, or 18.04 or for macOS: 10.14 or 10.15.'
        example: "20.04"
        in: query
        name: pv
        required: true
        type: string
      - description: Machine architecture, valid values are returned by the `/architectures`
          endpoint.
        example: x86_64
        in: query
        name: m
        required: true
        type: string
      - default: latest
        description: Version of the product to be installed. A version always takes
          the form `x.y.z`
        in: query
        name: v
        type: string
      - description: License ID
        in: query
        name: license_id
        type: string
      - default: false
        description: EOL Products
        in: query
        name: eol
        type: boolean
      produces:
      - application/json
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Download a product package
  /{channel}/{product}/fileName:
    get:
      consumes:
      - application/json
      description: The `ACCEPT` HTTP header with a value of `application/json` must
        be provided in the request for a JSON response to be returned
      parameters:
      - description: Channel
        enum:
        - current
        - stable
        in: path
        name: channel
        required: true
        type: string
      - description: Product
        example: chef
        in: path
        name: product
        required: true
        type: string
      - description: Platform, valid values are returned from the `/platforms` endpoint.
        example: ubuntu
        in: query
        name: p
        required: true
        type: string
      - description: 'Platform Version, possible values depend on the platform. For
          example, Ubuntu: 16.04, or 18.04 or for macOS: 10.14 or 10.15.'
        example: "20.04"
        in: query
        name: pv
        required: true
        type: string
      - description: Machine architecture, valid values are returned by the `/architectures`
          endpoint.
        example: x86_64
        in: query
        name: m
        required: true
        type: string
      - default: latest
        description: Version of the product to be installed. A version always takes
          the form `x.y.z`
        in: query
        name: v
        type: string
      - description: License ID
        in: query
        name: license_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get filename for a product package
  /{channel}/{product}/metadata:
    get:
      consumes:
      - application/json
      description: |-
        Get details for a particular package.
        The `ACCEPT` HTTP header with a value of `application/json` must be provided in the request for a JSON response to be returned
      parameters:
      - description: Channel
        enum:
        - current
        - stable
        in: path
        name: channel
        required: true
        type: string
      - description: Product
        example: chef
        in: path
        name: product
        required: true
        type: string
      - description: Platform, valid values are returned from the `/platforms` endpoint.
        example: ubuntu
        in: query
        name: p
        required: true
        type: string
      - description: 'Platform Version, possible values depend on the platform. For
          example, Ubuntu: 16.04, or 18.04 or for macOS: 10.14 or 10.15.'
        example: "20.04"
        in: query
        name: pv
        required: true
        type: string
      - description: Machine architecture, valid values are returned by the `/architectures`
          endpoint.
        example: x86_64
        in: query
        name: m
        required: true
        type: string
      - default: latest
        description: Version of the product to be installed. A version always takes
          the form `x.y.z`
        in: query
        name: v
        type: string
      - description: License ID
        in: query
        name: license_id
        type: string
      - default: false
        description: EOL Products
        in: query
        name: eol
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omnitruck.PackageMetadata'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get metadata for a product
  /{channel}/{product}/packages:
    get:
      consumes:
      - application/json
      description: |-
        Get the full list of all packages for a particular channel and product combination.
        By default all packages for the latest version are returned. If the v query string parameter is included the packages for the specified version are returned.
      parameters:
      - description: Channel
        enum:
        - current
        - stable
        in: path
        name: channel
        required: true
        type: string
      - description: Product
        example: chef
        in: path
        name: product
        required: true
        type: string
      - description: Version
        in: query
        name: v
        type: string
      - description: License ID
        in: query
        name: license_id
        type: string
      - default: false
        description: EOL Products
        in: query
        name: eol
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omnitruck.PackageList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get packages for a product version
  /{channel}/{product}/versions/all:
    get:
      consumes:
      - application/json
      description: Get a list of all available version numbers for a particular channel
        and product combination
      parameters:
      - description: Channel
        enum:
        - current
        - stable
        in: path
        name: channel
        required: true
        type: string
      - description: Product
        in: path
        name: product
        required: true
        type: string
      - description: License ID
        in: query
        name: license_id
        type: string
      - default: false
        description: EOL Products
        in: query
        name: eol
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get all versions of a product in a channel
  /{channel}/{product}/versions/latest:
    get:
      consumes:
      - application/json
      description: Get the latest version number for a particular channel and product
        combination.
      parameters:
      - description: Channel
        enum:
        - current
        - stable
        in: path
        name: channel
        required: true
        type: string
      - description: Product
        in: path
        name: product
        required: true
        type: string
      - description: License ID
        in: query
        name: license_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get latest version of a product in a channel
  /architectures:
    get:
      consumes:
      - application/json
      description: |-
        Returns a valid list of valid platform keys along with friendly names.
        Any of these architecture keys can be used in the p query string value in various endpoints below.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get architecture keys
  /install.ps1:
    get:
      consumes:
      - application/json
      description: The `ACCEPT` HTTP header with a value of `text/plain` must be provided
        in the request for a text response to be returned
      parameters:
      - description: License ID
        in: query
        name: license_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Download install script for Windows
  /install.sh:
    get:
      consumes:
      - application/json
      description: The `ACCEPT` HTTP header with a value of `application/x-sh` must
        be provided in the request for a shell script response to be returned
      parameters:
      - description: License ID
        in: query
        name: license_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Download install script for Linux
  /package-managers:
    get:
      consumes:
      - application/json
      description: Get the list of available package managers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get available package managers
...
swagger: "2.0"